+++ tests/unit_tests/test_docs/test-embed-hide-post-details-config/conf.py
@@ -1,0 +1,8 @@
+"""Sphinx test configuration."""
+exclude_patterns = ["_build"]
+extensions = ["sphinx_imgur.imgur"]
+html_theme = "basic"
+master_doc = "index"
+nitpicky = True
+
+imgur_hide_post_details = True